+++ CHANGELOG.md
@@ -837,6 +837,9 @@
 
 - Include refunds in the report CSV exports
 - BEIS users can export spending breakdown for all organisations
+- BEIS users can export spending breakdown for any organisation
+- Delviery partner users can export spending breakdown for their own
+  organisation
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-75...HEAD
 [release-75]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-74...release-75